[PROJ-67] refactor: chat reservation page css
